refactor: move HeavySwarm prompts and schemas to swarms/prompts/heavy_swarm_prompts.py
Extract all prompt constants and tool schemas from heavy_swarm.py into a
dedicated prompts file, reducing the struct file by ~490 lines while
keeping re-exports intact for backwards compatibility.
